[FIX]Uso de super para heredar valores desde modulos superirores
y que no se pierdan
