+++ sale_order_currency/models/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from . import sale
+from . import sale_order_line
+from . import sale_order
