--- sale_order_currency/models/sale.py
@@ -1,91 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from odoo import tools
-from odoo import api, fields, models
-import logging
-_logger = logging.getLogger(__name__)
-
-class SaleOrder(models.Model):
-    _inherit = ["sale.order"]
-    invoice_currency = fields.Boolean('Home Currency', default=True)
-
-    @api.multi
-    def _prepare_invoice(self):
-        """
-        Prepare the dict of values to create the new invoice for a sales order. This method may be
-        overridden to implement custom invoice generation (making sure to call super() to establish
-        a clean extension chain).
-        """
-        self.ensure_one()
-        journal_id = self.env['account.invoice'].default_get(['journal_id'])['journal_id']
-        if not journal_id:
-            raise UserError(_('Please define an accounting sales journal for this company.'))
-
-        if self.invoice_currency:
-            account_currency=self.env.user.company_id.currency_id.id
-            nombre=self.env.user.company_id.currency_id.name
-        else:
-            account_currency=self.pricelist_id.currency_id.id
-            nombre=self.pricelist_id.currency_id.name
-
-        invoice_vals = {
-            'name': self.client_order_ref or '',
-            'origin': self.name,
-            'type': 'out_invoice',
-            'account_id': self.partner_invoice_id.property_account_receivable_id.id,
-            'partner_id': self.partner_invoice_id.id,
-            'partner_shipping_id': self.partner_shipping_id.id,
-            'journal_id': journal_id,
-            'currency_id': account_currency,
-            'comment': self.note,
-            'payment_term_id': self.payment_term_id.id,
-            'fiscal_position_id': self.fiscal_position_id.id or self.partner_invoice_id.property_account_position_id.id,
-            'company_id': self.company_id.id,
-            'user_id': self.user_id and self.user_id.id,
-            'team_id': self.team_id.id
-        }
-        return invoice_vals
-
-
-
-class SaleOrderLine(models.Model):
-    _inherit = ["sale.order.line"]
-    @api.multi
-    def _prepare_invoice_line(self, qty):
-        """
-        Prepare the dict of values to create the new invoice line for a sales order line.
-        :param qty: float quantity to invoice
-        """
-        self.ensure_one()
-        res = {}
-        account = self.product_id.property_account_income_id or self.product_id.categ_id.property_account_income_categ_id
-        if not account:
-            raise UserError(_('Please define income account for this product: "%s" (id:%d) - or for its category: "%s".') %
-                (self.product_id.name, self.product_id.id, self.product_id.categ_id.name))
-
-        fpos = self.order_id.fiscal_position_id or self.order_id.partner_id.property_account_position_id
-        if fpos:
-            account = fpos.map_account(account)
-
-        if self.order_id.invoice_currency:
-            price_converted=self.price_unit / self.order_id.pricelist_id.currency_id.rate
-        else:
-            price_converted=self.price_unit
-
-        res = {
-            'name': self.name,
-            'sequence': self.sequence,
-            'origin': self.order_id.name,
-            'account_id': account.id,
-            'price_unit': price_converted,
-            'quantity': qty,
-            'discount': self.discount,
-            'uom_id': self.product_uom.id,
-            'product_id': self.product_id.id or False,
-            'layout_category_id': self.layout_category_id and self.layout_category_id.id or False,
-            'invoice_line_tax_ids': [(6, 0, self.tax_id.ids)],
-            'account_analytic_id': self.order_id.analytic_account_id.id,
-            'analytic_tag_ids': [(6, 0, self.analytic_tag_ids.ids)],
-        }
-        return res
