--- sale_order_currency/views/sale_views.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-        <!-- Top menu item -->
-        
-
-        <record id="view_order_form_currency" model="ir.ui.view">
-            <field name="name">sale.order.form</field>
-            <field name="model">sale.order</field>
-            <field name="inherit_id" ref="sale.view_order_form" />
-            <field name="arch" type="xml">
-                <field name="confirmation_date" position="after">
-                    <field name="invoice_currency"/>
-                </field>
-            </field>
-        </record>
-
-
-</odoo>
